refactor: backpack: move getURNWithoutTokenId to utils/URN-utils.ts
