+++ scene/src/ui-classes/main-menu/backpack-page/BackpackPage.tsx
@@ -10,9 +10,9 @@ import {WearableCatalogGrid} from "../../../components/backpack/WearableCatalogG
 import {
     catalystWearableMap,
     fetchWearablesData,
-    fetchWearablesPage,
-    getURNWithoutTokenId
+    fetchWearablesPage
 } from "../../../utils/wearables-promise-utils";
+
 import { getPlayer } from '@dcl/sdk/src/players'
 import type {URN, URNWithoutTokenId} from "../../../utils/definitions";
 import type {
@@ -31,6 +31,7 @@ import {
     setAvatarPreviewCameraToWearableCategory
 } from '../../../components/backpack/AvatarPreview'
 import {BASE_MALE_URN, ROUNDED_TEXTURE_BACKGROUND, ZERO_ADDRESS} from "../../../utils/constants";
+import {getURNWithoutTokenId} from "../../../utils/URN-utils";
 
 const WEARABLE_CATALOG_PAGE_SIZE = 16;
