+++ scene/src/utils/wearables-promise-utils.ts
@@ -85,13 +85,3 @@ export async function fetchWearablesData(...wearableURNs:URNWithoutTokenId[]):Pr
     }
 }
 
-const urnWithoutTokenIdMemo:Record<URN, URNWithoutTokenId> = {} // TODO consider using Map if possible for performance improvement because long keys
-
-export function getURNWithoutTokenId(urn:URN|null):URN|null{
-    if(urn === null) return null
-    if(urnWithoutTokenIdMemo[urn] !== undefined) return urnWithoutTokenIdMemo[urn];
-    // TODO add unit test?
-    // TODO should not break the URN when executed 2 times
-    urnWithoutTokenIdMemo[urn] = (urn.includes(":off-chain:") || urn.split(":").length < 6 ? urn : urn.replace(/^(.*):[^:]+$/, "$1")) as URNWithoutTokenId
-    return urnWithoutTokenIdMemo[urn];
-}
